6c/6l amd64: fix 64-bit constant truncation (long vs int32)
Two independent instances of this project's recurring long-vs-int32-
on-a-64-bit-host bug, both hit by any 64-bit constant (integer or IEEE
double literal) wider than 32 bits, e.g. NaN's 0x7FF0000000000001 or
plain doubles like 2.2:

- include/6.out.h's Ieee.l/.h fields were `long`. The linkers copy the
  raw bytes of this struct straight into the data segment (6l's AMOVSD/
  AFMOVD case in obj.c), so on our 64-bit hosts every IEEE double
  literal got corrupted down to its low 32 bits (with the high bits
  either zeroed or sign-extended garbage, depending on the low half).
  Already correct in the shared include/common.out.h (used by 5l/8l,
  with the exact same claude: comment from a previous session) and in
  the older linkers/8lk/6.out.h copy -- this freshly-imported vanilla
  header just still had the original 32-bit-host `long`.

- compilers/6c/swt.c's zaddr() used `long l` for its "does this offset
  fit in 32 bits" check (l = a->offset; if((vlong)l != a->offset) t |=
  T_64;), which can never be true when long is 64 bits, so the T_64
  object-file flag was never set and linkers/6l/obj.c's zaddr() (the
  reader) never read the high 32 bits back. Go-era's own 6g (GO/cmd/
  6g/gobj.c zaddr()) already uses int32 here for exactly this reason.

Also keeps linkers/6l/span.c's oclass() `long l` -> `int32 l` fix (same
"does this fit in 32 bits" pattern, matching Go-era src/cmd/6l/span.c):
real bug, though not what caused the symptom actually chased down here.

Adds tests/c/variants/fconst64_amd64.c as a minimal runnable regression
test (verified it fails against the pre-fix toolchain, passes after).

+++ compilers/6c/swt.c
@@ -360,7 +360,14 @@ zname(Biobuf *b, Sym *s, int t)
 void
 zaddr(Biobuf *b, Adr *a, int s)
 {
-	long l;
+	/* claude: was `long l`. On our 64-bit host `long` is 64 bits, so
+	 * `l = a->offset; if((vlong)l != a->offset)` never differs and
+	 * T_64 was never set, silently truncating any constant that
+	 * doesn't fit in 32 bits down to its low 32 bits. Go-era's own
+	 * 6g (GO/cmd/6g/gobj.c, zaddr()) already uses int32 here for
+	 * exactly this reason -- matching it.
+	 */
+	int32 l;
 	int i, t;
 	char *n;
 	Ieee e;
