+++ linkers/6l/span.c
@@ -363,7 +363,7 @@ int
 oclass(Adr *a)
 {
 	vlong v;
-	long l;
+	int32 l;
 
 	if(a->type >= D_INDIR || a->index != D_NONE) {
 		if(a->index != D_NONE && a->scale == 0) {
